Remove step mode
