+++ packages/react-native-sortables/src/constants/props.ts
@@ -23,12 +23,8 @@ export const DEFAULT_SHARED_PROPS = {
   autoScrollEnabled: true,
   autoScrollExtrapolation: Extrapolation.EXTEND,
   autoScrollInterval: { get: () => (IS_ANDROID && isFabric() ? 300 : 0) },
-  autoScrollMaxInterval: 500,
-  autoScrollMaxJump: 3,
   autoScrollMaxOverscroll: 50,
   autoScrollMaxVelocity: 1000,
-  autoScrollMinInterval: 100,
-  autoScrollMode: 'continuous',
   bringToFrontWhenActive: true,
   customHandle: false,
   debug: false,
