--- packages/react-native-sortables/src/providers/shared/AutoScrollProvider/modes/index.ts
@@ -1,2 +1,0 @@
-export { default as ContinuousModeUpdater } from './ContinuousModeUpdater';
-export { default as StepModeUpdater } from './StepModeUpdater';
